refactor Atlassian markdown converters
